+++ Data/PrecombatBuffs.lua
@@ -149,7 +149,8 @@ SpellDB.RegisterPrecombatBuffs({
         -- Dragonflight / TWW
         { id = 225592, buff = 455369 },  -- Exquisitely Eviscerated Muscle
         { id = 222740, buff = 382235, stat = "stamina" },  -- Hearty Meals
-        { id = 222745, buff = 462148 },  -- Pep-In-Your-Step
+        { id = 223971, buff = 456578, stat = "speed" },  -- Azj-Kahet Special
+        { id = 212492, buff = 454497, stat = "speed" },  -- Twilight-Spiced Grouper
         -- Shadowlands
         { id = 197783, buff = 382153, stat = "haste+versatility" },  -- Aromatic Seafood Platter
         { id = 197788, buff = 382234, stat = "stamina" },  -- Braised Bruffalon Brisket
@@ -204,7 +205,6 @@ SpellDB.RegisterPrecombatBuffs({
         { id = 154885, buff = 257422, stat = "versatility" },  -- Mon'Dazi
         { id = 154883, buff = 257413, stat = "haste" },  -- Ravenberry Tarts
         { id = 166343, buff = 288074, stat = "stamina" },  -- Wild Berry Bread
-        { id = 162117, buff = 274375, stat = "versatility" },  -- Marie's Famous Cookies
         { id = 167832, buff = 294365 },  -- Canned Minnows
         -- Warlords of Draenor
         { id = 133571, buff = 225603, stat = "haste" },  -- Azshari Salad
@@ -254,8 +254,6 @@ SpellDB.RegisterPrecombatBuffs({
         -- Wrath
         { id = 62654, buff = 87634, stat = "haste" },  -- Lavascale Fillet
         -- Classic
-        { id = 202401, buff = 19705, stat = "stamina" },  -- Cactus Apple Surprise
-        { id = 133576, buff = 201670 },  -- Bear Tartare
         { id = 143634, buff = 225597, stat = "crit" },  -- Critical Catfish
         { id = 143633, buff = 225599, stat = "mastery" },  -- Masterful Mana Buns
         { id = 111449, buff = 160889, stat = "crit" },  -- Blackrock Barbecue
@@ -290,6 +288,7 @@ SpellDB.RegisterPrecombatBuffs({
         { id = 98127, buff = 124219, stat = "mastery" },  -- Dented Can of Kaja'Cola
         { id = 101747, buff = 146807, stat = "versatility" },  -- Farmer's Delight
         { id = 101750, buff = 146804, stat = "strength" },  -- Fluffy Silkfeather Omelet
+        { id = 98116, buff = 142280, stat = "speed" },  -- Freeze-Dried Hyena Jerky
         { id = 104343, buff = 124216, stat = "haste" },  -- Golden Dragon Noodles
         { id = 94535, buff = 124219, stat = "mastery" },  -- Grilled Dinosaur Haunch
         { id = 79320, buff = 111840, stat = "stamina" },  -- Half a Lovely Apple
@@ -307,15 +306,11 @@ SpellDB.RegisterPrecombatBuffs({
         { id = 101749, buff = 146808, stat = "stamina" },  -- Stuffed Lushrooms
         { id = 81406, buff = 124213, stat = "mastery" },  -- Roasted Barley Tea
         { id = 86070, buff = 125106, stat = "crit" },  -- Wildfowl Ginseng Soup
-        { id = 103647, buff = 127770, stat = "haste" },  -- Chronambulatrix
         { id = 70924, buff = 100368, stat = "intellect" },  -- Eternal Eye of Elune
         { id = 70925, buff = 100373, stat = "agility" },  -- Eternal Lunar Pear
         { id = 70926, buff = 100375, stat = "versatility" },  -- Eternal Moonberry
         { id = 70927, buff = 100377, stat = "strength" },  -- Eternal Sunfruit
-        { id = 88382, buff = 127770, stat = "haste" },  -- Keenbean Kafa
         { id = 90457, buff = 131828, stat = "mastery" },  -- Mah's Warm Yak-Tail Stew
-        { id = 103646, buff = 127770, stat = "haste" },  -- Rejuvenating Potion
-        { id = 103645, buff = 127770, stat = "haste" },  -- Scroll of the Mystics
         { id = 89594, buff = 130356, stat = "intellect" },  -- Serpent Brew of Serenity
         { id = 89601, buff = 130348, stat = "agility" },  -- Tiger Brew of Serenity
         { id = 62661, buff = 87551, stat = "crit" },  -- Baked Rockfish
@@ -388,120 +383,41 @@ SpellDB.RegisterPrecombatBuffs({
         { id = 27651, buff = 33254, stat = "stamina" },  -- Buzzard Bites
         { id = 30155, buff = 33254, stat = "stamina" },  -- Clam Bar
         { id = 31673, buff = 33263, stat = "versatility" },  -- Crunchy Serpent
-        { id = 21023, buff = 25661, stat = "stamina" },  -- Dirge's Kickin' Chimaerok Chops
-        { id = 24009, buff = 145635, stat = "stamina" },  -- Dried Fruit Rations
-        { id = 24008, buff = 145635, stat = "stamina" },  -- Dried Mushroom Rations
         { id = 27662, buff = 33254, stat = "stamina" },  -- Feltail Delight
         { id = 27666, buff = 33268, stat = "versatility" },  -- Golden Fish Sticks
         { id = 27664, buff = 33261, stat = "agility" },  -- Grilled Mudfish
-        { id = 29292, buff = 24799, stat = "strength" },  -- Helboar Bacon
         { id = 140342, buff = 35272, stat = "stamina" },  -- Lean - Mok'Nathal Shortribs
-        { id = 24539, buff = 18193, stat = "versatility" },  -- Marsh Lichen
         { id = 31672, buff = 35272, stat = "stamina" },  -- Mok'Nathal Shortribs
         { id = 27665, buff = 33263, stat = "versatility" },  -- Poached Bluefish
         { id = 27655, buff = 33259 },  -- Ravager Dog
-        { id = 28501, buff = 145635, stat = "stamina" },  -- Ravager Egg Omelet
         { id = 27658, buff = 33256, stat = "strength" },  -- Roasted Clefthoof
         { id = 89590, buff = 130352, stat = "intellect" },  -- Serpent Brew of Adversity
-        { id = 32721, buff = 42293, stat = "stamina" },  -- Skyguard Rations
         { id = 27667, buff = 33257, stat = "stamina" },  -- Spicy Crawdad
         { id = 27660, buff = 35272, stat = "stamina" },  -- Talbuk Steak
         { id = 89597, buff = 130344, stat = "agility" },  -- Tiger Brew of Adversity
         { id = 27659, buff = 33261, stat = "agility" },  -- Warp Burger
-        { id = 13810, buff = 18125, stat = "strength" },  -- Blessed Sunfruit
-        { id = 35563, buff = 46899 },  -- Charred Bear Kabobs
         { id = 89589, buff = 130351, stat = "intellect" },  -- Initiate's Serpent Brew
         { id = 89596, buff = 130343, stat = "agility" },  -- Initiate's Tiger Brew
-        { id = 35565, buff = 46687, stat = "intellect" },  -- Juicy Bear Burger
-        { id = 13934, buff = 145635, stat = "stamina" },  -- Mightfish Steak
-        { id = 18254, buff = 22730, stat = "intellect" },  -- Runn Tum Tuber Surprise
-        { id = 20452, buff = 24799, stat = "strength" },  -- Smoked Desert Dumplings
-        { id = 16971, buff = 19710, stat = "stamina" },  -- Clamlette Surprise
-        { id = 12218, buff = 19710, stat = "stamina" },  -- Monster Omelet
-        { id = 12216, buff = 19710, stat = "stamina" },  -- Spiced Chili Crab
-        { id = 18045, buff = 19710, stat = "stamina" },  -- Tender Wolf Steak
         { id = 33004, buff = 64057 },  -- Clamlette Magnifique
-        { id = 13927, buff = 18191, stat = "stamina" },  -- Cooked Glossy Mightfish
-        { id = 13928, buff = 18192, stat = "agility" },  -- Grilled Squid
-        { id = 12215, buff = 19710, stat = "stamina" },  -- Heavy Kodo Stew
-        { id = 13929, buff = 19709, stat = "stamina" },  -- Hot Smoked Bass
         { id = 89588, buff = 130350, stat = "intellect" },  -- Novice's Serpent Brew
         { id = 89595, buff = 130342, stat = "agility" },  -- Novice's Tiger Brew
-        { id = 17222, buff = 19710, stat = "stamina" },  -- Spider Sausage
-        { id = 57519, buff = 19705, stat = "stamina" },  -- Cookie's Special Ramlette
-        { id = 4457, buff = 19709, stat = "stamina" },  -- Barbecued Buzzard Wing
-        { id = 12213, buff = 19709, stat = "stamina" },  -- Carrion Surprise
-        { id = 6038, buff = 19709, stat = "stamina" },  -- Giant Clam Scorcho
-        { id = 20074, buff = 19709, stat = "stamina" },  -- Heavy Crocolisk Stew
-        { id = 13851, buff = 19709, stat = "stamina" },  -- Hot Wolf Ribs
-        { id = 12212, buff = 19709, stat = "stamina" },  -- Jungle Stew
-        { id = 12214, buff = 19709, stat = "stamina" },  -- Mystery Stew
-        { id = 12210, buff = 19709, stat = "stamina" },  -- Roast Raptor
-        { id = 3729, buff = 19709, stat = "stamina" },  -- Soothing Turtle Bisque
-        { id = 3728, buff = 19709, stat = "stamina" },  -- Tasty Lion Steak
-        { id = 5342, buff = 6114, stat = "intellect" },  -- Raptor Punch
-        { id = 3726, buff = 19708, stat = "stamina" },  -- Big Bear Steak
-        { id = 5479, buff = 19708, stat = "stamina" },  -- Crispy Lizard Tail
-        { id = 3664, buff = 19708, stat = "stamina" },  -- Crocolisk Gumbo
-        { id = 3665, buff = 19708, stat = "stamina" },  -- Curiously Tasty Omelet
-        { id = 5527, buff = 19708, stat = "stamina" },  -- Goblin Deviled Clams
-        { id = 3666, buff = 19708, stat = "stamina" },  -- Gooey Spider Cake
-        { id = 3727, buff = 19708, stat = "stamina" },  -- Hot Lion Chops
-        { id = 5480, buff = 19708, stat = "stamina" },  -- Lean Venison
-        { id = 12209, buff = 19708, stat = "stamina" },  -- Lean Wolf Steak
-        { id = 3663, buff = 19708, stat = "stamina" },  -- Murloc Fin Soup
-        { id = 1082, buff = 19708, stat = "stamina" },  -- Redridge Goulash
-        { id = 1017, buff = 19708, stat = "stamina" },  -- Seasoned Wolf Kabob
-        { id = 27635, buff = 19705, stat = "stamina" },  -- Lynx Steak
-        { id = 24105, buff = 19705, stat = "stamina" },  -- Roasted Moongraze Tenderloin
-        { id = 27636, buff = 19706, stat = "stamina" },  -- Bat Bites
-        { id = 3220, buff = 19706, stat = "stamina" },  -- Blood Sausage
-        { id = 5525, buff = 19706, stat = "stamina" },  -- Boiled Clams
-        { id = 11584, buff = 19705, stat = "stamina" },  -- Cactus Apple Surprise
-        { id = 276611, buff = 19706, stat = "stamina" },  -- Clam Stuffed Crab
-        { id = 2684, buff = 19706, stat = "stamina" },  -- Coyote Steak
-        { id = 2683, buff = 19706, stat = "stamina" },  -- Crab Cake
-        { id = 3662, buff = 19706, stat = "stamina" },  -- Crocolisk Steak
-        { id = 22645, buff = 19706, stat = "stamina" },  -- Crunchy Spider Surprise
-        { id = 2687, buff = 19706, stat = "stamina" },  -- Dry Pork Ribs
-        { id = 5476, buff = 19706, stat = "stamina" },  -- Fillet of Frenzy
-        { id = 724, buff = 19706, stat = "stamina" },  -- Goretusk Liver Pie
-        { id = 140338, buff = 19706, stat = "stamina" },  -- Mana-Wrapped Goretusk Liver Pie
-        { id = 5477, buff = 19706, stat = "stamina" },  -- Strider Stew
-        { id = 2888, buff = 19705, stat = "stamina" },  -- Beer Basted Boar Ribs
-        { id = 23756, buff = 19705, stat = "stamina" },  -- Cookie's Jumbo Gumbo
-        { id = 12224, buff = 19705, stat = "stamina" },  -- Crispy Bat Wing
         { id = 184682, buff = 347173, stat = "versatility" },  -- Extra Lemony Herb Filet
-        { id = 6888, buff = 19705, stat = "stamina" },  -- Herb Baked Egg
-        { id = 5472, buff = 19705, stat = "stamina" },  -- Kaldorei Spider Kabob
         { id = 126934, buff = 185736, stat = "versatility" },  -- Lemon Herb Filet
-        { id = 5474, buff = 19705, stat = "stamina" },  -- Roasted Kodo Meat
-        { id = 77264, buff = 108028, stat = "intellect" },  -- Small Bamboo Shoot
-        { id = 77273, buff = 108031, stat = "strength" },  -- Small Rice Cake
-        { id = 77272, buff = 108032, stat = "agility" },  -- Small Sugarcane Stalk
-        { id = 2680, buff = 19705, stat = "stamina" },  -- Spiced Wolf Meat
         -- exp -3
-        { id = 210821, buff = 19711, stat = "stamina" },  -- Berry Chocolate Tart
-        { id = 210819, buff = 19710, stat = "stamina" },  -- Chocolate-Covered Cherries
         { id = 34411, buff = 45245, stat = "stamina" },  -- Hot Apple Cider
         { id = 21254, buff = 110645, stat = "primary" },  -- Winter Veil Cookie
         { id = 210842, buff = 226807, stat = "versatility" },  -- Fresh Fig
         { id = 210841, buff = 226805, stat = "mastery" },  -- Buttered Oysters
         { id = 44839, buff = 65410, stat = "haste" },  -- Candied Sweet Potato
-        { id = 7807, buff = 19705, stat = "stamina" },  -- Candy Bar
-        { id = 7808, buff = 19705, stat = "stamina" },  -- Chocolate Square
         { id = 116405, buff = 65412, stat = "stamina" },  -- Congealed Cranberry Chutney
         { id = 44840, buff = 226807, stat = "versatility" },  -- Cranberry Chutney
-        { id = 17197, buff = 19705, stat = "stamina" },  -- Gingerbread Cookie
         { id = 116409, buff = 175785, stat = "crit" },  -- Gourmet Pumpkin Pie
         { id = 116408, buff = 175784, stat = "stamina" },  -- Herb-Infused Stuffing
-        { id = 7806, buff = 19705, stat = "stamina" },  -- Lollipop
         { id = 44836, buff = 175785, stat = "crit" },  -- Pumpkin Pie
         { id = 44838, buff = 226805, stat = "mastery" },  -- Slow-Roasted Turkey
         { id = 116407, buff = 65414 },  -- Slow-Smoked Turkey
         { id = 44837, buff = 175784, stat = "stamina" },  -- Spice Bread Stuffing
         { id = 116406, buff = 65410, stat = "haste" },  -- Twice-Baked Sweet Potato
-        { id = 17198, buff = 19705, stat = "stamina" },  -- Winter Veil Eggnog
     },
     augmentRune = {
         -- Midnight / current
@@ -529,200 +445,199 @@ SpellDB.RegisterPrecombatBuffs({
     },
     weaponEnchant = {
         -- Midnight / current
-        { id = 243736, buff = 1237001 },  -- Oil of Dawn
-        { id = 237369, buff = 1224333 },  -- Refulgent Weightstone
-        { id = 237371, buff = 1224328 },  -- Refulgent Whetstone
-        { id = 243734, buff = 1237006 },  -- Thalassian Phoenix Oil
-        { id = 243735, buff = 1237002 },  -- Oil of Dawn
-        { id = 237367, buff = 1224332 },  -- Refulgent Weightstone
-        { id = 237370, buff = 1224331 },  -- Refulgent Whetstone
-        { id = 243733, buff = 1237008 },  -- Thalassian Phoenix Oil
+        { id = 243736, buff = 1237001, stat = "caster" },  -- Oil of Dawn
+        { id = 237369, buff = 1224333, stat = "physical" },  -- Refulgent Weightstone
+        { id = 237371, buff = 1224328, stat = "physical" },  -- Refulgent Whetstone
+        { id = 243734, buff = 1237006, stat = "caster" },  -- Thalassian Phoenix Oil
+        { id = 243735, buff = 1237002, stat = "caster" },  -- Oil of Dawn
+        { id = 237367, buff = 1224332, stat = "physical" },  -- Refulgent Weightstone
+        { id = 237370, buff = 1224331, stat = "physical" },  -- Refulgent Whetstone
+        { id = 243733, buff = 1237008, stat = "caster" },  -- Thalassian Phoenix Oil
         -- Dragonflight / TWW
-        { id = 220156, buff = 444755 },  -- Bubbling Wax
-        { id = 222510, buff = 458937 },  -- Ironclaw Weightstone
-        { id = 222504, buff = 458934 },  -- Ironclaw Whetstone
-        { id = 222509, buff = 458936 },  -- Ironclaw Weightstone
-        { id = 222503, buff = 458933 },  -- Ironclaw Whetstone
-        { id = 222508, buff = 458935 },  -- Ironclaw Weightstone
-        { id = 222502, buff = 458932 },  -- Ironclaw Whetstone
-        { id = 224105, buff = 451869 },  -- Algari Mana Oil
-        { id = 224106, buff = 451873 },  -- Algari Mana Oil
-        { id = 224107, buff = 451874 },  -- Algari Mana Oil
-        { id = 224108, buff = 451926 },  -- Oil of Beledar's Grace
-        { id = 224109, buff = 451925 },  -- Oil of Beledar's Grace
-        { id = 224110, buff = 451927 },  -- Oil of Beledar's Grace
-        { id = 224111, buff = 451882 },  -- Oil of Deep Toxins
-        { id = 224112, buff = 451901 },  -- Oil of Deep Toxins
-        { id = 224113, buff = 451902 },  -- Oil of Deep Toxins
-        { id = 234382, buff = 1213839 },  -- Ensemble: Unkindled Waxweave Panoply
-        { id = 226224, buff = 458517 },  -- Luminous Nerubian Catalyst Stone
+        { id = 220156, buff = 444755, stat = "caster" },  -- Bubbling Wax
+        { id = 222510, buff = 458937, stat = "physical" },  -- Ironclaw Weightstone
+        { id = 222504, buff = 458934, stat = "physical" },  -- Ironclaw Whetstone
+        { id = 222509, buff = 458936, stat = "physical" },  -- Ironclaw Weightstone
+        { id = 222503, buff = 458933, stat = "physical" },  -- Ironclaw Whetstone
+        { id = 222508, buff = 458935, stat = "physical" },  -- Ironclaw Weightstone
+        { id = 222502, buff = 458932, stat = "physical" },  -- Ironclaw Whetstone
+        { id = 224105, buff = 451869, stat = "caster" },  -- Algari Mana Oil
+        { id = 224106, buff = 451873, stat = "caster" },  -- Algari Mana Oil
+        { id = 224107, buff = 451874, stat = "caster" },  -- Algari Mana Oil
+        { id = 224108, buff = 451926, stat = "caster" },  -- Oil of Beledar's Grace
+        { id = 224109, buff = 451925, stat = "caster" },  -- Oil of Beledar's Grace
+        { id = 224110, buff = 451927, stat = "caster" },  -- Oil of Beledar's Grace
+        { id = 224111, buff = 451882, stat = "caster" },  -- Oil of Deep Toxins
+        { id = 224112, buff = 451901, stat = "caster" },  -- Oil of Deep Toxins
+        { id = 224113, buff = 451902, stat = "caster" },  -- Oil of Deep Toxins
+        { id = 234382, buff = 1213839, stat = "caster" },  -- Ensemble: Unkindled Waxweave Panoply
+        { id = 226224, buff = 458517, stat = "physical" },  -- Luminous Nerubian Catalyst Stone
         -- Shadowlands
-        { id = 191944, buff = 371677 },  -- Primal Weightstone
-        { id = 191940, buff = 396157 },  -- Primal Whetstone
-        { id = 191945, buff = 371678 },  -- Primal Weightstone
-        { id = 191933, buff = 396155 },  -- Primal Whetstone
-        { id = 191943, buff = 371676 },  -- Primal Weightstone
-        { id = 191939, buff = 396156 },  -- Primal Whetstone
+        { id = 191944, buff = 371677, stat = "physical" },  -- Primal Weightstone
+        { id = 191940, buff = 396157, stat = "physical" },  -- Primal Whetstone
+        { id = 191945, buff = 371678, stat = "physical" },  -- Primal Weightstone
+        { id = 191933, buff = 396155, stat = "physical" },  -- Primal Whetstone
+        { id = 191943, buff = 371676, stat = "physical" },  -- Primal Weightstone
+        { id = 191939, buff = 396156, stat = "physical" },  -- Primal Whetstone
         -- Battle for Azeroth
-        { id = 171436, buff = 322749 },  -- Porous Sharpening Stone
-        { id = 171438, buff = 322761 },  -- Porous Weightstone
-        { id = 171437, buff = 322762 },  -- Shaded Sharpening Stone
-        { id = 171439, buff = 322763 },  -- Shaded Weightstone
-        { id = 180683, buff = 334428 },  -- Guardian Stonewing
-        { id = 180684, buff = 334431 },  -- Sentinel Stonewing
-        { id = 181337, buff = 336160 },  -- Stonefiend's Wings
-        { id = 176139, buff = 321651 },  -- Contract: Stoneborn Reserves
-        { id = 186981, buff = 355159 },  -- Porous Stone Statue
-        { id = 178686, buff = 327299 },  -- RSVP: Stonehead
-        { id = 186982, buff = 355175 },  -- Shaded Stone Statue
-        { id = 181771, buff = 343719 },  -- Stonehead's Best Bud Bracelet
-        { id = 189572, buff = 365941 },  -- Tame Prime: Hadeon the Stonebreaker
+        { id = 171436, buff = 322749, stat = "physical" },  -- Porous Sharpening Stone
+        { id = 171438, buff = 322761, stat = "physical" },  -- Porous Weightstone
+        { id = 171437, buff = 322762, stat = "physical" },  -- Shaded Sharpening Stone
+        { id = 171439, buff = 322763, stat = "physical" },  -- Shaded Weightstone
+        { id = 180683, buff = 334428, stat = "physical" },  -- Guardian Stonewing
+        { id = 180684, buff = 334431, stat = "physical" },  -- Sentinel Stonewing
+        { id = 181337, buff = 336160, stat = "physical" },  -- Stonefiend's Wings
+        { id = 176139, buff = 321651, stat = "physical" },  -- Contract: Stoneborn Reserves
+        { id = 186981, buff = 355159, stat = "physical" },  -- Porous Stone Statue
+        { id = 178686, buff = 327299, stat = "physical" },  -- RSVP: Stonehead
+        { id = 186982, buff = 355175, stat = "physical" },  -- Shaded Stone Statue
+        { id = 181771, buff = 343719, stat = "physical" },  -- Stonehead's Best Bud Bracelet
+        { id = 189572, buff = 365941, stat = "physical" },  -- Tame Prime: Hadeon the Stonebreaker
         -- Legion
-        { id = 165719, buff = 285840 },  -- Monelite Whetstone
-        { id = 165720, buff = 285840 },  -- Monelite Whetstone
-        { id = 162120, buff = 276228 },  -- Platinum Whetstone
-        { id = 165748, buff = 276228 },  -- Platinum Whetstone
-        { id = 167077, buff = 300539 },  -- Scrying Stone
-        { id = 161385, buff = 275199 },  -- Stonefist Watch
-        { id = 162511, buff = 276927 },  -- Stonetusk Watch
+        { id = 165719, buff = 285840, stat = "physical" },  -- Monelite Whetstone
+        { id = 165720, buff = 285840, stat = "physical" },  -- Monelite Whetstone
+        { id = 162120, buff = 276228, stat = "physical" },  -- Platinum Whetstone
+        { id = 165748, buff = 276228, stat = "physical" },  -- Platinum Whetstone
+        { id = 167077, buff = 300539, stat = "physical" },  -- Scrying Stone
+        { id = 161385, buff = 275199, stat = "physical" },  -- Stonefist Watch
+        { id = 162511, buff = 276927, stat = "physical" },  -- Stonetusk Watch
         -- Warlords of Draenor
-        { id = 147303, buff = 240515 },  -- Giant Elemental's Closed Stone Fist
-        { id = 152927, buff = 253837 },  -- Serrated Stone Axe
-        { id = 131746, buff = 208705 },  -- Stonehide Leather Barding
+        { id = 147303, buff = 240515, stat = "physical" },  -- Giant Elemental's Closed Stone Fist
+        { id = 152927, buff = 253837, stat = "physical" },  -- Serrated Stone Axe
+        { id = 131746, buff = 208705, stat = "physical" },  -- Stonehide Leather Barding
         -- Mists of Pandaria
-        { id = 118221, buff = 173893 },  -- Petrification Stone
+        { id = 118221, buff = 173893, stat = "physical" },  -- Petrification Stone
         -- Cataclysm
-        { id = 87872, buff = 127563 },  -- Desecrated Oil
+        { id = 87872, buff = 127563, stat = "caster" },  -- Desecrated Oil
         -- Classic
-        { id = 23529, buff = 29453 },  -- Adamantite Sharpening Stone
-        { id = 28421, buff = 34340 },  -- Adamantite Weightstone
-        { id = 22522, buff = 28017 },  -- Superior Wizard Oil
-        { id = 23528, buff = 29452 },  -- Fel Sharpening Stone
-        { id = 28420, buff = 34339 },  -- Fel Weightstone
-        { id = 25884, buff = 32806 },  -- Primal Stone Statue
-        { id = 22521, buff = 28013 },  -- Superior Mana Oil
-        { id = 29796, buff = 35745 },  -- Socrethar's Teleportation Stone
+        { id = 23529, buff = 29453, stat = "physical" },  -- Adamantite Sharpening Stone
+        { id = 28421, buff = 34340, stat = "physical" },  -- Adamantite Weightstone
+        { id = 22522, buff = 28017, stat = "caster" },  -- Superior Wizard Oil
+        { id = 23528, buff = 29452, stat = "physical" },  -- Fel Sharpening Stone
+        { id = 28420, buff = 34339, stat = "physical" },  -- Fel Weightstone
+        { id = 25884, buff = 32806, stat = "physical" },  -- Primal Stone Statue
+        { id = 22521, buff = 28013, stat = "caster" },  -- Superior Mana Oil
+        { id = 29796, buff = 35745, stat = "physical" },  -- Socrethar's Teleportation Stone
         -- Classic
-        { id = 140329, buff = 224382 },  -- Infinite Stone
-        { id = 146919, buff = 240307 },  -- Portal-Stone: An'thyna
-        { id = 146920, buff = 240310 },  -- Portal-Stone: Fel Obliterator
-        { id = 146918, buff = 240304 },  -- Portal-Stone: Force-Commander Xillious
-        { id = 146921, buff = 240313 },  -- Portal-Stone: Illisthyndria
-        { id = 146917, buff = 240301 },  -- Portal-Stone: Skulguloth
-        { id = 146916, buff = 240298 },  -- Portal-Stone: Than'otalion
-        { id = 116981, buff = 172376 },  -- Fire Ammonite Oil
-        { id = 122457, buff = 181062 },  -- Ultimate Battle-Training Stone
-        { id = 133890, buff = 201800 },  -- Stoneshroom
-        { id = 116119, buff = 170895 },  -- Ango'rosh Sorcerer Stone
-        { id = 62345, buff = 86901 },  -- Greater Inscription of Jagged Stone
-        { id = 68716, buff = 86901 },  -- Greater Inscription of Jagged Stone
-        { id = 62344, buff = 86900 },  -- Lesser Inscription of Jagged Stone
-        { id = 89303, buff = 129894 },  -- Stack of Stone Blocks
-        { id = 36899, buff = 47904 },  -- Exceptional Mana Oil
-        { id = 18262, buff = 22756 },  -- Elemental Sharpening Stone
-        { id = 23123, buff = 28898 },  -- Blessed Wizard Oil
-        { id = 203862, buff = 25123 },  -- Brilliant Mana Oil
-        { id = 203865, buff = 25122 },  -- Brilliant Wizard Oil
-        { id = 23122, buff = 28891 },  -- Consecrated Sharpening Stone
-        { id = 25883, buff = 32805 },  -- Dense Stone Statue
-        { id = 20747, buff = 25120 },  -- Lesser Mana Oil
-        { id = 20750, buff = 25121 },  -- Wizard Oil
-        { id = 12404, buff = 16138 },  -- Dense Sharpening Stone
-        { id = 12643, buff = 16622 },  -- Dense Weightstone
-        { id = 8956, buff = 11350 },  -- Oil of Immolation
-        { id = 118699, buff = 11350 },  -- Oil of Immolation
-        { id = 3829, buff = 3595 },  -- Frost Oil
-        { id = 20746, buff = 25119 },  -- Lesser Wizard Oil
-        { id = 25882, buff = 32804 },  -- Solid Stone Statue
-        { id = 7964, buff = 9900 },  -- Solid Sharpening Stone
-        { id = 7965, buff = 9903 },  -- Solid Weightstone
-        { id = 3824, buff = 3594 },  -- Shadow Oil
-        { id = 25881, buff = 32803 },  -- Heavy Stone Statue
-        { id = 20745, buff = 25118 },  -- Minor Mana Oil
-        { id = 2871, buff = 2830 },  -- Heavy Sharpening Stone
-        { id = 3241, buff = 3114 },  -- Heavy Weightstone
-        { id = 25880, buff = 32802 },  -- Coarse Stone Statue
-        { id = 1322, buff = 7396, stat = "haste" },  -- Fishliver Oil
-        { id = 2863, buff = 2829 },  -- Coarse Sharpening Stone
-        { id = 3240, buff = 3113 },  -- Coarse Weightstone
-        { id = 20744, buff = 25117 },  -- Minor Wizard Oil
-        { id = 6458, buff = 433 },  -- Oil Covered Fish
-        { id = 25498, buff = 32254 },  -- Rough Stone Statue
-        { id = 1177, buff = 54497 },  -- Oil of Olaf
-        { id = 1434, buff = 13424 },  -- Glowing Wax Stick
-        { id = 2862, buff = 2828 },  -- Rough Sharpening Stone
-        { id = 3239, buff = 3112 },  -- Rough Weightstone
-        { id = 116424, buff = 171515 },  -- Aquatic Battle-Training Stone
-        { id = 87257, buff = 126589 },  -- Arcane Familiar Stone
-        { id = 116374, buff = 162984 },  -- Beast Battle-Training Stone
-        { id = 116418, buff = 171520 },  -- Critter Battle-Training Stone
-        { id = 147564, buff = 242134 },  -- Diamond Stone
-        { id = 116419, buff = 171516 },  -- Dragonkin Battle-Training Stone
-        { id = 116420, buff = 171522 },  -- Elemental Battle-Training Stone
-        { id = 127755, buff = 187567 },  -- Fel-Touched Battle-Training Stone
-        { id = 87258, buff = 126587 },  -- Fiery Familiar Stone
-        { id = 92679, buff = 134488 },  -- Flawless Aquatic Battle-Stone
-        { id = 92741, buff = 134644 },  -- Flawless Battle-Stone
-        { id = 116429, buff = 171525 },  -- Flawless Battle-Training Stone
-        { id = 92675, buff = 134482 },  -- Flawless Beast Battle-Stone
-        { id = 92676, buff = 134484 },  -- Flawless Critter Battle-Stone
-        { id = 92683, buff = 134492 },  -- Flawless Dragonkin Battle-Stone
-        { id = 92665, buff = 134448 },  -- Flawless Elemental Battle-Stone
-        { id = 92677, buff = 134486 },  -- Flawless Flying Battle-Stone
-        { id = 92682, buff = 134491 },  -- Flawless Humanoid Battle-Stone
-        { id = 92678, buff = 134487 },  -- Flawless Magic Battle-Stone
-        { id = 92680, buff = 134489 },  -- Flawless Mechanical Battle-Stone
-        { id = 92681, buff = 134490 },  -- Flawless Undead Battle-Stone
-        { id = 116421, buff = 171518 },  -- Flying Battle-Training Stone
-        { id = 8095, buff = 10723 },  -- Hinott's Oil
-        { id = 116416, buff = 171517 },  -- Humanoid Battle-Training Stone
-        { id = 87259, buff = 126586 },  -- Icy Familiar Stone
-        { id = 137391, buff = 212863 },  -- Immaculate Aquatic Battle-Stone
-        { id = 137627, buff = 213521 },  -- Immaculate Battle-Stone
-        { id = 137394, buff = 212862 },  -- Immaculate Beast Battle-Stone
-        { id = 137393, buff = 212856 },  -- Immaculate Critter Battle-Stone
-        { id = 137387, buff = 212859 },  -- Immaculate Dragonkin Battle-Stone
-        { id = 137395, buff = 212857 },  -- Immaculate Elemental Battle-Stone
-        { id = 137396, buff = 212860 },  -- Immaculate Flying Battle-Stone
-        { id = 137388, buff = 212866 },  -- Immaculate Humanoid Battle-Stone
-        { id = 137392, buff = 212858 },  -- Immaculate Magic Battle-Stone
-        { id = 137390, buff = 212861 },  -- Immaculate Mechanical Battle-Stone
-        { id = 137389, buff = 212864 },  -- Immaculate Undead Battle-Stone
-        { id = 142359, buff = 230286 },  -- Lava Oil
-        { id = 116422, buff = 171521 },  -- Magic Battle-Training Stone
-        { id = 98715, buff = 134644 },  -- Marked Flawless Battle-Stone
-        { id = 116417, buff = 171523 },  -- Mechanical Battle-Training Stone
-        { id = 92689, buff = 134499 },  -- Polished Aquatic Battle-Stone
-        { id = 92742, buff = 134496 },  -- Polished Battle-Stone
-        { id = 92685, buff = 134495 },  -- Polished Beast Battle-Stone
-        { id = 92686, buff = 134761 },  -- Polished Critter Battle-Stone
-        { id = 92693, buff = 134504 },  -- Polished Dragonkin Battle-Stone
-        { id = 92684, buff = 134494 },  -- Polished Elemental Battle-Stone
-        { id = 92687, buff = 134497 },  -- Polished Flying Battle-Stone
-        { id = 92692, buff = 134503 },  -- Polished Humanoid Battle-Stone
-        { id = 92688, buff = 134498 },  -- Polished Magic Battle-Stone
-        { id = 92690, buff = 134501 },  -- Polished Mechanical Battle-Stone
-        { id = 92691, buff = 134502 },  -- Polished Undead Battle-Stone
-        { id = 219146, buff = 442935 },  -- Set Keystone Map: Stonevault
-        { id = 147563, buff = 242133 },  -- Stone Totem
-        { id = 30309, buff = 11009 },  -- Stonebreaker Brew
-        { id = 116423, buff = 171519 },  -- Undead Battle-Training Stone
-        { id = 139816, buff = 221923 },  -- Well-Worn Stone
+        { id = 140329, buff = 224382, stat = "physical" },  -- Infinite Stone
+        { id = 146919, buff = 240307, stat = "physical" },  -- Portal-Stone: An'thyna
+        { id = 146920, buff = 240310, stat = "physical" },  -- Portal-Stone: Fel Obliterator
+        { id = 146918, buff = 240304, stat = "physical" },  -- Portal-Stone: Force-Commander Xillious
+        { id = 146921, buff = 240313, stat = "physical" },  -- Portal-Stone: Illisthyndria
+        { id = 146917, buff = 240301, stat = "physical" },  -- Portal-Stone: Skulguloth
+        { id = 146916, buff = 240298, stat = "physical" },  -- Portal-Stone: Than'otalion
+        { id = 116981, buff = 172376, stat = "caster" },  -- Fire Ammonite Oil
+        { id = 122457, buff = 181062, stat = "physical" },  -- Ultimate Battle-Training Stone
+        { id = 133890, buff = 201800, stat = "physical" },  -- Stoneshroom
+        { id = 116119, buff = 170895, stat = "physical" },  -- Ango'rosh Sorcerer Stone
+        { id = 62345, buff = 86901, stat = "physical" },  -- Greater Inscription of Jagged Stone
+        { id = 68716, buff = 86901, stat = "physical" },  -- Greater Inscription of Jagged Stone
+        { id = 62344, buff = 86900, stat = "physical" },  -- Lesser Inscription of Jagged Stone
+        { id = 89303, buff = 129894, stat = "physical" },  -- Stack of Stone Blocks
+        { id = 36899, buff = 47904, stat = "caster" },  -- Exceptional Mana Oil
+        { id = 18262, buff = 22756, stat = "physical" },  -- Elemental Sharpening Stone
+        { id = 23123, buff = 28898, stat = "caster" },  -- Blessed Wizard Oil
+        { id = 203862, buff = 25123, stat = "caster" },  -- Brilliant Mana Oil
+        { id = 203865, buff = 25122, stat = "caster" },  -- Brilliant Wizard Oil
+        { id = 23122, buff = 28891, stat = "physical" },  -- Consecrated Sharpening Stone
+        { id = 25883, buff = 32805, stat = "physical" },  -- Dense Stone Statue
+        { id = 20747, buff = 25120, stat = "caster" },  -- Lesser Mana Oil
+        { id = 20750, buff = 25121, stat = "caster" },  -- Wizard Oil
+        { id = 12404, buff = 16138, stat = "physical" },  -- Dense Sharpening Stone
+        { id = 12643, buff = 16622, stat = "physical" },  -- Dense Weightstone
+        { id = 8956, buff = 11350, stat = "caster" },  -- Oil of Immolation
+        { id = 118699, buff = 11350, stat = "caster" },  -- Oil of Immolation
+        { id = 3829, buff = 3595, stat = "caster" },  -- Frost Oil
+        { id = 20746, buff = 25119, stat = "caster" },  -- Lesser Wizard Oil
+        { id = 25882, buff = 32804, stat = "physical" },  -- Solid Stone Statue
+        { id = 7964, buff = 9900, stat = "physical" },  -- Solid Sharpening Stone
+        { id = 7965, buff = 9903, stat = "physical" },  -- Solid Weightstone
+        { id = 3824, buff = 3594, stat = "caster" },  -- Shadow Oil
+        { id = 25881, buff = 32803, stat = "physical" },  -- Heavy Stone Statue
+        { id = 20745, buff = 25118, stat = "caster" },  -- Minor Mana Oil
+        { id = 2871, buff = 2830, stat = "physical" },  -- Heavy Sharpening Stone
+        { id = 3241, buff = 3114, stat = "physical" },  -- Heavy Weightstone
+        { id = 25880, buff = 32802, stat = "physical" },  -- Coarse Stone Statue
+        { id = 1322, buff = 7396, stat = "caster" },  -- Fishliver Oil
+        { id = 2863, buff = 2829, stat = "physical" },  -- Coarse Sharpening Stone
+        { id = 3240, buff = 3113, stat = "physical" },  -- Coarse Weightstone
+        { id = 20744, buff = 25117, stat = "caster" },  -- Minor Wizard Oil
+        { id = 6458, buff = 433, stat = "caster" },  -- Oil Covered Fish
+        { id = 25498, buff = 32254, stat = "physical" },  -- Rough Stone Statue
+        { id = 1177, buff = 54497, stat = "caster" },  -- Oil of Olaf
+        { id = 1434, buff = 13424, stat = "caster" },  -- Glowing Wax Stick
+        { id = 2862, buff = 2828, stat = "physical" },  -- Rough Sharpening Stone
+        { id = 3239, buff = 3112, stat = "physical" },  -- Rough Weightstone
+        { id = 116424, buff = 171515, stat = "physical" },  -- Aquatic Battle-Training Stone
+        { id = 87257, buff = 126589, stat = "physical" },  -- Arcane Familiar Stone
+        { id = 116374, buff = 162984, stat = "physical" },  -- Beast Battle-Training Stone
+        { id = 116418, buff = 171520, stat = "physical" },  -- Critter Battle-Training Stone
+        { id = 147564, buff = 242134, stat = "physical" },  -- Diamond Stone
+        { id = 116419, buff = 171516, stat = "physical" },  -- Dragonkin Battle-Training Stone
+        { id = 116420, buff = 171522, stat = "physical" },  -- Elemental Battle-Training Stone
+        { id = 127755, buff = 187567, stat = "physical" },  -- Fel-Touched Battle-Training Stone
+        { id = 87258, buff = 126587, stat = "physical" },  -- Fiery Familiar Stone
+        { id = 92679, buff = 134488, stat = "physical" },  -- Flawless Aquatic Battle-Stone
+        { id = 92741, buff = 134644, stat = "physical" },  -- Flawless Battle-Stone
+        { id = 116429, buff = 171525, stat = "physical" },  -- Flawless Battle-Training Stone
+        { id = 92675, buff = 134482, stat = "physical" },  -- Flawless Beast Battle-Stone
+        { id = 92676, buff = 134484, stat = "physical" },  -- Flawless Critter Battle-Stone
+        { id = 92683, buff = 134492, stat = "physical" },  -- Flawless Dragonkin Battle-Stone
+        { id = 92665, buff = 134448, stat = "physical" },  -- Flawless Elemental Battle-Stone
+        { id = 92677, buff = 134486, stat = "physical" },  -- Flawless Flying Battle-Stone
+        { id = 92682, buff = 134491, stat = "physical" },  -- Flawless Humanoid Battle-Stone
+        { id = 92678, buff = 134487, stat = "physical" },  -- Flawless Magic Battle-Stone
+        { id = 92680, buff = 134489, stat = "physical" },  -- Flawless Mechanical Battle-Stone
+        { id = 92681, buff = 134490, stat = "physical" },  -- Flawless Undead Battle-Stone
+        { id = 116421, buff = 171518, stat = "physical" },  -- Flying Battle-Training Stone
+        { id = 8095, buff = 10723, stat = "caster" },  -- Hinott's Oil
+        { id = 116416, buff = 171517, stat = "physical" },  -- Humanoid Battle-Training Stone
+        { id = 87259, buff = 126586, stat = "physical" },  -- Icy Familiar Stone
+        { id = 137391, buff = 212863, stat = "physical" },  -- Immaculate Aquatic Battle-Stone
+        { id = 137627, buff = 213521, stat = "physical" },  -- Immaculate Battle-Stone
+        { id = 137394, buff = 212862, stat = "physical" },  -- Immaculate Beast Battle-Stone
+        { id = 137393, buff = 212856, stat = "physical" },  -- Immaculate Critter Battle-Stone
+        { id = 137387, buff = 212859, stat = "physical" },  -- Immaculate Dragonkin Battle-Stone
+        { id = 137395, buff = 212857, stat = "physical" },  -- Immaculate Elemental Battle-Stone
+        { id = 137396, buff = 212860, stat = "physical" },  -- Immaculate Flying Battle-Stone
+        { id = 137388, buff = 212866, stat = "physical" },  -- Immaculate Humanoid Battle-Stone
+        { id = 137392, buff = 212858, stat = "physical" },  -- Immaculate Magic Battle-Stone
+        { id = 137390, buff = 212861, stat = "physical" },  -- Immaculate Mechanical Battle-Stone
+        { id = 137389, buff = 212864, stat = "physical" },  -- Immaculate Undead Battle-Stone
+        { id = 142359, buff = 230286, stat = "caster" },  -- Lava Oil
+        { id = 116422, buff = 171521, stat = "physical" },  -- Magic Battle-Training Stone
+        { id = 98715, buff = 134644, stat = "physical" },  -- Marked Flawless Battle-Stone
+        { id = 116417, buff = 171523, stat = "physical" },  -- Mechanical Battle-Training Stone
+        { id = 92689, buff = 134499, stat = "physical" },  -- Polished Aquatic Battle-Stone
+        { id = 92742, buff = 134496, stat = "physical" },  -- Polished Battle-Stone
+        { id = 92685, buff = 134495, stat = "physical" },  -- Polished Beast Battle-Stone
+        { id = 92686, buff = 134761, stat = "physical" },  -- Polished Critter Battle-Stone
+        { id = 92693, buff = 134504, stat = "physical" },  -- Polished Dragonkin Battle-Stone
+        { id = 92684, buff = 134494, stat = "physical" },  -- Polished Elemental Battle-Stone
+        { id = 92687, buff = 134497, stat = "physical" },  -- Polished Flying Battle-Stone
+        { id = 92692, buff = 134503, stat = "physical" },  -- Polished Humanoid Battle-Stone
+        { id = 92688, buff = 134498, stat = "physical" },  -- Polished Magic Battle-Stone
+        { id = 92690, buff = 134501, stat = "physical" },  -- Polished Mechanical Battle-Stone
+        { id = 92691, buff = 134502, stat = "physical" },  -- Polished Undead Battle-Stone
+        { id = 219146, buff = 442935, stat = "physical" },  -- Set Keystone Map: Stonevault
+        { id = 147563, buff = 242133, stat = "physical" },  -- Stone Totem
+        { id = 30309, buff = 11009, stat = "physical" },  -- Stonebreaker Brew
+        { id = 116423, buff = 171519, stat = "physical" },  -- Undead Battle-Training Stone
+        { id = 139816, buff = 221923, stat = "physical" },  -- Well-Worn Stone
         -- exp -3
-        { id = 223647, buff = 449639 },  -- Lifeless Stone Ring
+        { id = 223647, buff = 449639, stat = "physical" },  -- Lifeless Stone Ring
     },
 })
 
 -- ──────────────────────────── HAND-CURATED (not generated) ────────────────────────────
--- Toys and class self-buffs flow through the same detect+overlay pipeline via `source`.
--- Add them with SpellDB.RegisterPrecombatBuffsExtra so a regen never clobbers them.
+-- Curated extras the per-class generator above doesn't cover: aura-discovered utility
+-- categories (xp / movement-speed foods, duration-filtered to 20m+) plus toys and class
+-- self-buffs. All flow through the same detect+overlay pipeline via `source`. This block is
+-- part of the generator template, so a re-run preserves it verbatim - edit it here.
 if SpellDB.RegisterPrecombatBuffsExtra then
     SpellDB.RegisterPrecombatBuffsExtra({
-        -- Speed (movement): long well-fed FOODS only (>= 20m buff) - a maintained pre-buff,
-        -- not a 15s potion burst (useless as an OOC reminder). Off by default.
-        { category = "speed", id = 212492, buff = 454497 },   -- Twilight-Spiced Grouper (60m)
-        { category = "speed", id = 223971, buff = 456578 },   -- Azj-Kahet Special (20m)
-        { category = "speed", id = 98116,  buff = 142280 },   -- Freeze-Dried Hyena Jerky (60m)
+        -- (Movement-speed foods are generated into the `food` category above, tagged
+        --  stat = "speed" - they share the one Well Fed slot, so they're a food option.)
 
         -- XP (leveling): long XP buffs only (>= 20m). Off by default. Tome of Combat Training
         -- (10m) intentionally excluded by the duration floor.